hash_fst {fst}	R Documentation

Parallel calculation of the hash of a raw vector

Description

Parallel calculation of the hash of a raw vector

Usage

hash_fst(x, seed = NULL, block_hash = TRUE)

Arguments

x
raw vector that you want to hash

seed
The seed value for the hashing algorithm. If NULL, a default seed will be used.

block_hash
If TRUE, a multi-threaded implementation of the 64-bit xxHash algorithm will be used. Note that block_hash = TRUE or block_hash = FALSE will result in different hash values.

Value

hash value

[Package fst version 0.9.4 Index]