metadata_fst {fst}	R Documentation

Read metadata from a fst file

Description

Method for checking basic properties of the dataset stored in path.

Usage

metadata_fst(path, old_format = FALSE)

fst.metadata(path, old_format = FALSE)

Arguments

path
path to fst file

old_format
must be FALSE, the old fst file format is deprecated and can only be read and converted with fst package versions 0.8.0 to 0.8.10.

Value

Returns a list with meta information on the stored dataset in path. Has class fstmetadata.

Examples

# Sample dataset
x <- data.frame(
  First = 1:10,
  Second = sample(c(TRUE, FALSE, NA), 10, replace = TRUE),
  Last = sample(LETTERS, 10))

# Write to fst file
fst_file <- tempfile(fileext = ".fst")
write_fst(x, fst_file)

# Display meta information
metadata_fst(fst_file)

[Package fst version 0.9.2 Index]