fst-package {fst}	R Documentation

Lightning Fast Serialization of Data Frames for R.

Description

Multithreaded serialization of compressed data frames using the 'fst' format. The 'fst' format allows for random access of stored data which can be compressed with the LZ4 and ZSTD compressors.

Details

The fst package is based on three C++ libraries:

fstlib: library containing code to write, read and compute on files stored in the fst format. Written and owned by Mark Klik.

LZ4: library containing code to compress data with the LZ4 compressor. Written and owned by Yann Collet.

ZSTD: library containing code to compress data with the ZSTD compressor. Written by Yann Collet and owned by Facebook, Inc.

As of version 0.9.8, these libraries are included in the fstcore package, on which fst depends. The copyright notices of the above libraries can be found in the fstcore package.

[Package fst version 0.9.8 Index]